Use appropriate types for syscall arguments. NFC
The net effect of this change simply that there are fewer i64 argument
in the syscall fuctions for wasm64.  For wasm32 this change has no
effect at all.

Unlike the linux syscall layer we are not bount to fixed interface
consisting only of word-sized (long) arguments.

For wasm64 we want to avoid the use of `long` in order to limit the
number of BitInt arguments on the JS side.  In this change I've modified
the arguments from all being `long` into three different classes:

- `int`: For 32-bit values - these are the same under
         wasm32 and wasm64.
- `intptr_t/size_t`: For pointer-sized things that are
                     64-bit only under wasm64.
- `uint64_t`: For `off_t` and similar that are always
              64-bit.

There is very little usage of always-64-bit/off_t in the syscall layer
and I already landed that part of the change in: #16590

+++ system/lib/wasmfs/syscalls.cpp
@@ -44,7 +44,7 @@ extern "C" {
 
 using namespace wasmfs;
 
-int __syscall_dup3(long oldfd, long newfd, long flags) {
+int __syscall_dup3(int oldfd, int newfd, int flags) {
   if (flags & !O_CLOEXEC) {
     // TODO: Test this case.
     return -EINVAL;
@@ -68,7 +68,7 @@ int __syscall_dup3(long oldfd, long newfd, long flags) {
   return newfd;
 }
 
-int __syscall_dup(long fd) {
+int __syscall_dup(int fd) {
   auto fileTable = wasmFS.getFileTable().locked();
 
   // Check that an open file exists corresponding to the given fd.
@@ -303,7 +303,7 @@ backend_t wasmfs_get_backend_by_path(const char* path) {
   return parsed.getFile()->getBackend();
 }
 
-int __syscall_fstatat64(long dirfd, long path, long buf, long flags) {
+int __syscall_fstatat64(int dirfd, intptr_t path, intptr_t buf, int flags) {
   // Only accept valid flags.
   if (flags & ~(AT_EMPTY_PATH | AT_NO_AUTOMOUNT | AT_SYMLINK_NOFOLLOW)) {
     // TODO: Test this case.
@@ -362,20 +362,20 @@ int __syscall_fstatat64(long dirfd, long path, long buf, long flags) {
   return __WASI_ERRNO_SUCCESS;
 }
 
-int __syscall_stat64(long path, long buf) {
+int __syscall_stat64(intptr_t path, intptr_t buf) {
   return __syscall_fstatat64(AT_FDCWD, path, buf, 0);
 }
 
-int __syscall_lstat64(long path, long buf) {
+int __syscall_lstat64(intptr_t path, intptr_t buf) {
   return __syscall_fstatat64(AT_FDCWD, path, buf, AT_SYMLINK_NOFOLLOW);
 }
 
-int __syscall_fstat64(long fd, long buf) {
-  return __syscall_fstatat64(fd, (long)"", buf, AT_EMPTY_PATH);
+int __syscall_fstat64(int fd, intptr_t buf) {
+  return __syscall_fstatat64(fd, (intptr_t)"", buf, AT_EMPTY_PATH);
 }
 
 static __wasi_fd_t doOpen(path::ParsedParent parsed,
-                          long flags,
+                          int flags,
                           mode_t mode,
                           backend_t backend = NullBackend) {
   int accessMode = (flags & O_ACCMODE);
@@ -467,7 +467,7 @@ int wasmfs_create_file(char* pathname, mode_t mode, backend_t backend) {
 }
 
 // TODO: Test this with non-AT_FDCWD values.
-int __syscall_openat(long dirfd, long path, long flags, ...) {
+int __syscall_openat(int dirfd, intptr_t path, int flags, ...) {
   mode_t mode = 0;
   va_list v1;
   va_start(v1, flags);
@@ -477,8 +477,8 @@ int __syscall_openat(long dirfd, long path, long flags, ...) {
   return doOpen(path::parseParent((char*)path, dirfd), flags, mode);
 }
 
-static long
-doMkdir(path::ParsedParent parsed, long mode, backend_t backend = NullBackend) {
+static int
+doMkdir(path::ParsedParent parsed, int mode, backend_t backend = NullBackend) {
   if (auto err = parsed.getError()) {
     return err;
   }
@@ -525,18 +525,18 @@ doMkdir(path::ParsedParent parsed, long mode, backend_t backend = NullBackend) {
 
 // This function is exposed to users and allows users to specify a particular
 // backend that a directory should be created within.
-int wasmfs_create_directory(char* path, long mode, backend_t backend) {
-  static_assert(std::is_same_v<decltype(doMkdir(0, 0, 0)), long>,
+int wasmfs_create_directory(char* path, int mode, backend_t backend) {
+  static_assert(std::is_same_v<decltype(doMkdir(0, 0, 0)), int>,
                 "unexpected conversion from result of doMkdir to int");
   return doMkdir(path::parseParent(path), mode, backend);
 }
 
 // TODO: Test this.
-int __syscall_mkdirat(long dirfd, long path, long mode) {
+int __syscall_mkdirat(int dirfd, intptr_t path, int mode) {
   return doMkdir(path::parseParent((char*)path, dirfd), mode);
 }
 
-int __syscall_mkdir(long path, long mode) {
+int __syscall_mkdir(intptr_t path, int mode) {
   return doMkdir(path::parseParent((char*)path), mode);
 }
 
@@ -580,7 +580,7 @@ __wasi_errno_t __wasi_fd_seek(__wasi_fd_t fd,
   return __WASI_ERRNO_SUCCESS;
 }
 
-long doChdir(std::shared_ptr<File>& file) {
+int doChdir(std::shared_ptr<File>& file) {
   auto dir = file->dynCast<Directory>();
   if (!dir) {
     return -ENOTDIR;
@@ -589,7 +589,7 @@ long doChdir(std::shared_ptr<File>& file) {
   return 0;
 }
 
-int __syscall_chdir(long path) {
+int __syscall_chdir(intptr_t path) {
   auto parsed = path::parseFile((char*)path);
   if (auto err = parsed.getError()) {
     return err;
@@ -597,7 +597,7 @@ int __syscall_chdir(long path) {
   return doChdir(parsed.getFile());
 }
 
-int __syscall_fchdir(long fd) {
+int __syscall_fchdir(int fd) {
   auto openFile = wasmFS.getFileTable().locked().getEntry(fd);
   if (!openFile) {
     return -EBADF;
@@ -605,7 +605,7 @@ int __syscall_fchdir(long fd) {
   return doChdir(openFile->locked().getFile());
 }
 
-int __syscall_getcwd(long buf, long size) {
+int __syscall_getcwd(intptr_t buf, size_t size) {
   // Check if buf points to a bad address.
   if (!buf && size > 0) {
     return -EFAULT;
@@ -674,7 +674,7 @@ __wasi_errno_t __wasi_fd_fdstat_get(__wasi_fd_t fd, __wasi_fdstat_t* stat) {
 }
 
 // TODO: Test this with non-AT_FDCWD values.
-int __syscall_unlinkat(long dirfd, long path, long flags) {
+int __syscall_unlinkat(int dirfd, intptr_t path, int flags) {
   if (flags & ~AT_REMOVEDIR) {
     // TODO: Test this case.
     return -EINVAL;
@@ -736,11 +736,11 @@ int __syscall_unlinkat(long dirfd, long path, long flags) {
   return 0;
 }
 
-int __syscall_rmdir(long path) {
+int __syscall_rmdir(intptr_t path) {
   return __syscall_unlinkat(AT_FDCWD, path, AT_REMOVEDIR);
 }
 
-int __syscall_getdents64(long fd, long dirp, long count) {
+int __syscall_getdents64(int fd, intptr_t dirp, size_t count) {
   dirent* result = (dirent*)dirp;
 
   // Check if the result buffer is too small.
@@ -813,10 +813,10 @@ int __syscall_getdents64(long fd, long dirp, long count) {
 }
 
 // TODO: Test this with non-AT_FDCWD values.
-int __syscall_renameat(long olddirfd,
-                       long oldpath,
-                       long newdirfd,
-                       long newpath) {
+int __syscall_renameat(int olddirfd,
+                       intptr_t oldpath,
+                       int newdirfd,
+                       intptr_t newpath) {
   // Rename is the only syscall that needs to (or is allowed to) acquire locks
   // on two directories at once. It requires locks on both the old and new
   // parent directories to ensure that the moved file can be atomically removed
@@ -924,12 +924,12 @@ int __syscall_renameat(long olddirfd,
   return 0;
 }
 
-int __syscall_rename(long oldpath, long newpath) {
+int __syscall_rename(intptr_t oldpath, intptr_t newpath) {
   return __syscall_renameat(AT_FDCWD, oldpath, AT_FDCWD, newpath);
 }
 
 // TODO: Test this with non-AT_FDCWD values.
-int __syscall_symlinkat(long target, long newdirfd, long linkpath) {
+int __syscall_symlinkat(intptr_t target, int newdirfd, intptr_t linkpath) {
   auto parsed = path::parseParent((char*)linkpath, newdirfd);
   if (auto err = parsed.getError()) {
     return err;
@@ -953,12 +953,12 @@ int __syscall_symlinkat(long target, long newdirfd, long linkpath) {
   return 0;
 }
 
-int __syscall_symlink(long target, long linkpath) {
+int __syscall_symlink(intptr_t target, intptr_t linkpath) {
   return __syscall_symlinkat(target, AT_FDCWD, linkpath);
 }
 
 // TODO: Test this with non-AT_FDCWD values.
-int __syscall_readlinkat(long dirfd, long path, long buf, long bufsize) {
+int __syscall_readlinkat(int dirfd, intptr_t path, intptr_t buf, size_t bufsize) {
   auto parsed = path::parseFile((char*)path, dirfd);
   if (auto err = parsed.getError()) {
     return err;
@@ -974,7 +974,7 @@ int __syscall_readlinkat(long dirfd, long path, long buf, long bufsize) {
 }
 
 // TODO: Test this with non-AT_FDCWD values.
-int __syscall_utimensat(long dirFD, long path_, long times_, long flags) {
+int __syscall_utimensat(int dirFD, intptr_t path_, intptr_t times_, int flags) {
   const char* path = (const char*)path_;
   const struct timespec* times = (const struct timespec*)times_;
   if (flags & ~AT_SYMLINK_NOFOLLOW) {
@@ -1008,7 +1008,7 @@ int __syscall_utimensat(long dirFD, long path_, long times_, long flags) {
 }
 
 // TODO: Test this with non-AT_FDCWD values.
-int __syscall_fchmodat(long dirfd, long path, long mode, ...) {
+int __syscall_fchmodat(int dirfd, intptr_t path, int mode, ...) {
   int flags = 0;
   va_list v1;
   va_start(v1, mode);
@@ -1028,12 +1028,12 @@ int __syscall_fchmodat(long dirfd, long path, long mode, ...) {
   return 0;
 }
 
-int __syscall_chmod(long path, long mode) {
+int __syscall_chmod(intptr_t path, int mode) {
   return __syscall_fchmodat(AT_FDCWD, path, mode, 0);
 }
 
 // TODO: Test this with non-AT_FDCWD values.
-int __syscall_faccessat(long dirfd, long path, long amode, long flags) {
+int __syscall_faccessat(int dirfd, intptr_t path, int amode, int flags) {
   // The input must be F_OK (check for existence) or a combination of [RWX]_OK
   // flags.
   if (amode != F_OK && (amode & ~(R_OK | W_OK | X_OK))) {
@@ -1066,7 +1066,7 @@ int __syscall_faccessat(long dirfd, long path, long amode, long flags) {
   return 0;
 }
 
-static long doTruncate(std::shared_ptr<File>& file, off_t size) {
+static int doTruncate(std::shared_ptr<File>& file, off_t size) {
   auto dataFile = file->dynCast<DataFile>();
   // TODO: support for symlinks.
   if (!dataFile) {
@@ -1089,7 +1089,7 @@ static long doTruncate(std::shared_ptr<File>& file, off_t size) {
   return 0;
 }
 
-int __syscall_truncate64(long path, uint64_t size) {
+int __syscall_truncate64(intptr_t path, uint64_t size) {
   auto parsed = path::parseFile((char*)path);
   if (auto err = parsed.getError()) {
     return err;
@@ -1097,7 +1097,7 @@ int __syscall_truncate64(long path, uint64_t size) {
   return doTruncate(parsed.getFile(), size);
 }
 
-int __syscall_ftruncate64(long fd, uint64_t size) {
+int __syscall_ftruncate64(int fd, uint64_t size) {
   auto openFile = wasmFS.getFileTable().locked().getEntry(fd);
   if (!openFile) {
     return -EBADF;
@@ -1112,7 +1112,7 @@ int __syscall_ftruncate64(long fd, uint64_t size) {
   return ret;
 }
 
-int __syscall_pipe(long fd) {
+int __syscall_pipe(intptr_t fd) {
   auto* fds = (__wasi_fd_t*)fd;
 
   // Make a pipe: Two PipeFiles that share a single data source between them, so
@@ -1133,13 +1133,13 @@ int __syscall_pipe(long fd) {
 }
 
 // int poll(struct pollfd* fds, nfds_t nfds, int timeout);
-int __syscall_poll(long fds_, long nfds, long timeout) {
+int __syscall_poll(intptr_t fds_, int nfds, int timeout) {
   struct pollfd* fds = (struct pollfd*)fds_;
   auto fileTable = wasmFS.getFileTable().locked();
 
   // Process the list of FDs and compute their revents masks. Count the number
   // of nonzero such masks, which is our return value.
-  long nonzero = 0;
+  int nonzero = 0;
   for (nfds_t i = 0; i < nfds; i++) {
     auto* pollfd = &fds[i];
     auto fd = pollfd->fd;
